requirements.txt
dotted_dict
google-api-core==1.31.6
google-api-python-client==2.28.0
google-apitools==0.5.31
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.6.2
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-bigquery==2.30.1
google-cloud-bigquery-storage==2.9.1
google-cloud-bigtable==1.7.2
google-cloud-core==1.7.3
google-cloud-dataproc==3.1.1
google-cloud-datastore==1.15.5
google-cloud-dlp==3.7.1
google-cloud-firestore==2.3.4
google-cloud-kms==2.10.1
google-cloud-language==1.3.2
google-cloud-logging==2.7.0
google-cloud-monitoring==2.6.0
google-cloud-pubsub==2.13.1
google-cloud-pubsublite==1.4.2
google-cloud-recommendations-ai==0.2.0
google-cloud-scheduler==2.5.0
google-cloud-spanner==1.19.3
google-cloud-speech==2.11.1
google-cloud-storage==1.42.3
google-cloud-tasks==2.7.1
google-cloud-translate==3.6.1
google-cloud-videointelligence==1.16.3
google-cloud-vision==1.0.2
google-crc32c==1.1.2
google-resumable-media==2.1.0
googleapis-common-protos==1.56.4
apache-beam==2.40.0